CHROMIUM: Arm: Enable state switches GPIO.
This change defines the three GPIOs which are connected to the
developer and recovery mode switches and the 'write protect'
signal, which indicates that the part of the firmware storage is
in read only mode.

The GPIOs statuses get exposed through /sys/class/gpio and are
used by the crossystem utility to represent the target state in a
consistent cross platform manner.

The new definitions match Kaen schematics. This is hardcoded
currently, but eventually would be derived from the device tree
supplied by the bootloader.

BUG=chromium-os:12522
TEST=manual:

. build and boot up the new kernel on a Kaen system
. run the following commands:
 localhost ~ # egrep -w '(56|59|168)' /sys/kernel/debug/gpio
 gpio-56  (recovery_switch     ) in  hi
 gpio-59  (wp_status           ) in  lo
 gpio-168 (dev_switch          ) in  lo
 localhost ~ # ls /sys/class/gpio/
 export  gpio160  gpio168  gpio185  gpio56  gpio59  gpiochip0  gpiochip224  gpiochip228  unexport
 localhost ~ #
. run command

 localhost ~ # echo $(crossystem recoverysw_cur)

 with the recovery button pressed and released, observe the change in the
 reported signal level.

Change-Id: I59b6f676e708083710b36a861efa8227dd63d7ff
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/418
Reviewed-by: Olof Johansson <olofj@chromium.org>