power: bq20z75: support external notification
Adding support for external power change notification. One problem found
is that there is a lag time before the sensor will return a new status. To
ensure that we only fire off the power_supply_changed event when the status
returned from the sensor is actually different, we delay sending the
the notification, and instead poll on it looking for a change. The amount
of time to poll is configurable via platform data.

TEST=unplug ac and see kernel event is fired off when change is detected,
same with plugging in. If status doesn't change in the specified time limit,
then no event is fired.
BUG=chrome-os-partner:2467

Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Change-Id: If895293c8bbc5c84751573ecd3c3090f0c9bfc5a
Reviewed-on: http://gerrit.chromium.org/gerrit/869
Reviewed-by: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>