CHROMIUM: ARM: tegra: arthur: Set the panel's bits per pixel to 32.
If the panel's pixel depth is set to 24, the Tegra video driver silently
errors out during part of its initialization and leaves some fields of some
descriptive structures uninitialized. This change sets the bits per pixel to
32, a value that's handled properly.

It also increase the amount of backing memory allocated for the frame buffer
to 0x97f680, enough space for a 1368x910 resolution, 32 bits per pixel, and
double buffering.

Finally, it also increases the bits per pixel sent to the panel to 24.

BUG=chrome-os-partner:3648
TEST=Built and installed the kernel on an existing system, ran the tests
described in the bug, and saw that the crashes were gone.

Signed-off-by: Gabe Black <gabeblack@google.com>
Change-Id: I27d85e2c74b48b3439211313c7b20c617eb9aacb
Reviewed-on: http://gerrit.chromium.org/gerrit/910
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>