net: rfkill: add generic gpio rfkill driver
This adds a new generic gpio rfkill driver to support rfkill switches
which are controlled by gpios. The driver also supports passing in
data about the clock for the radio, so that when rfkill is blocking,
it can disable the clock.

This driver assumes platform data is passed from the board files to
configure it for specific devices.

TEST=verified when enabled this driver can control BT radio and clock
BUG=chrome-os-partner:1949

Change-Id: I07d58f2c5515c012a61dc3b78603c05fad7a722a
Original-patch-by: Anantha Idapalapati <aidapalapati@nvidia.com>
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Signed-off-by: John W. Linville <linville@tuxdriver.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/1496
Reviewed-by: Olof Johansson <olofj@chromium.org>