CHROMIUM: ARM: tegra: arthur: Turn down the SDIO frequency.
This change temporarily turns down the SDIO frequency on Arthur until the
brcmfmac stability issues are resolved. This change makes it possible to
communicate with the wireless hardware, but a different NVRAM image needs to
be installed to it to be able to establish a connection.

BUG=chrome-os-partner:3579
TEST=Built and installed the kernel on an existing system, verified that
the wireless card could scan for access points.

Change-Id: I6f6661989d450431162826722e35e566b9cb35aa
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/997
Reviewed-by: Doug Anderson <dianders%chromium.org@gtempaccount.com>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>