CHROMIUM: arm: tegra: seaboard: Enable HDMI +5V for hotplug
This change enables the HDMI +5V GPIO when hotplug is necessary, and
disables it on suspend.

It turns out that some HDMI sinks actually need +5V all the time, since
they rely on it to provide the hotplug-detect current back to the
source.  So we can't turn it on only when doing DDC or TMDS traffic, but
need to leave it on all the time.

Signed-off-by: Robert Morell <rmorell@nvidia.com>

BUG=cros-partner:1812
TEST=HDMI hotplug detect on Tegra2 Seaboard-based platforms, verify that
+5V is not asserted in suspend
Change-Id: Ic1504968abb54e3e09a8242e87df94b34e2ad58b
Reviewed-on: http://gerrit.chromium.org/gerrit/1095
Reviewed-by: Olof Johansson <olofj@chromium.org>
Tested-by: Olof Johansson <olofj@chromium.org>
[olofj: removed warning due to type mismatch]
Signed-off-by: Olof Johansson <olof@lixom.net>