CHROMIUM: video: tegra: Share the display between dc_ext and fbdev
This change adds code to transition between fbdev and dc_ext modes of
operation.  Initially, the display is in fbdev mode.  Right before the
first dc_ext window is flipped to, the fbdev surface is removed.  When
the last dc_ext window is disabled, the fbdev surface is reinstated.

This allows the FB console to coexist with an accelerated X server, and
transition between the two via VT switches.

Signed-off-by: Robert Morell <rmorell@nvidia.com>

BUG=chromium-os-partner:1670
TEST=Verify that the FB console works, and that VT switching with an X
server present works.
Change-Id: I6ac7c53ae811522ec2c89ab1896771c571047034
Reviewed-on: http://gerrit.chromium.org/gerrit/1589
Tested-by: Robert Morell <rmorell@nvidia.com>
Reviewed-by: Olof Johansson <olofj@chromium.org>