CHROMIUM: ARM: tegra: kaen: reset CPU/TPM via GPIO
We are changing tegra2 kaen boards so that asserting
a GPIO resets the CPU and the TPM at the same time.  This also
tries not to break existing (prototype) hardware.

The restart flow is similar to the old one, except that
we don't turn off caching (cpu_proc_fin) and we don't do
the final cache flush.  Turning off the cache prevents
gpio_direction_output from working, and a lower-level
interface is not exposed.  Let me know if this isn't
acceptable.

TESTING.  I tested this by verifying that "reboot" works
on old and new hardware, and the expected kernel
messages are printed.  I could not test that RAM
is preserved because a log spew is preventing running the
logging_KernelCrashServer autotest.  However, frh has checked the
reset and power lines with a scope and everything works as
expected.

BUG=chromeos-partner:3574
TEST=verify that "reboot" command works properly, and that kernel crash logging still works.

Signed-off-by: Luigi Semenzato <semenzato@chromium.org>

Change-Id: I639fb21452a49eba5d5cb15fd120082f6a5c30c5
Reviewed-on: http://gerrit.chromium.org/gerrit/1920
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>