CHROMIUM: arm: tegra: arthur: Initial board port
This is a starter CL for something that we think will boot to a command prompt.

BUG=chrome-os-partner:3472
TEST=Kernel compiled and still boots on a different ARM board to validate that it didn't break with this.

NOTE: I also ran a smoke test on the ARM board that I have.  It got failures on 'suite_Smoke/login_LoginSuccess.apps' and 'suite_Smoke/login_LoginSuccess.default', but I don't think those are related to this change--it may be that I have an old source tree checked out.

Review URL: http://codereview.chromium.org/6905051

Change-Id: If9e7630e6683f02e934e52d8d5c6440f68be2443
Signed-off-by: Doug Anderson <dianders@chromium.org>