CHROMIUM: arm: tegra: harmony: Fix backlight toggling on fb0
In addition to the internal panel (on fb0), Harmony may have external
displays such as HDMI (on fb1).  One of the devices might be blanked
while another is unblanked.  In order for the backlight blanking logic
to know whether to turn off the backlight for a particular framebuffer's
blanking notification, it needs to be able to check if a given
framebuffer device corresponds to the backlight.

This implement the check_fb pwm_backlight hook to make the internal
panel's backlight only toggle on fb0 blank changes.

Signed-off-by: Robert Morell <rmorell@nvidia.com>

Change-Id: I29b613c02d4276ee10922dce6c09921d4644e587
BUG=none
TEST=HDMI on Harmony, powering off HDMI does not also power off internal
backlight
Reviewed-on: http://gerrit.chromium.org/gerrit/2115
Tested-by: Robert Morell <rmorell@nvidia.com>
Reviewed-by: Olof Johansson <olofj@chromium.org>