CHROMIUM: init: add root=PARTUUID=UUID/PARTNROFF=%d support
[sent upstream, still under review v3]

Expand root=PARTUUID=UUID syntax to support selecting a root partition
by integer offset from a known, unique partition.  This approach
provides similar properties to specifying a device and partition number,
but using the UUID as the unique path prior to evaluating the offset.

For example,
  root=PARTUUID=99DE9194-FC15-4223-9192-FC243948F88B/PARTNROFF=1
selects the partition with UUID 99DE.. then select the next
partition.

This change is motivated by a particular usecase in Chromium OS where
the bootloader can easily determine what partition it is on (by UUID)
but doesn't perform general partition table walking.

That said, support for this model provides a direct mechanism for the
user to modify the root partition to boot without specifically needing
to extract each UUID or update the bootloader explicitly when the root
partition UUID is changed (if it is recreated to be larger, for
instance).  Pinning to a /boot-style partition UUID allows the arbitrary
root partition reconfiguration/modifications with slightly less
ambiguity than just [dev][partition] and less stringency than the
specific root partition UUID.

At present, PARTNROFF is tied to PARTUUID use and must follow
immediately.  It is possible to expand support to other values of root=
or to add other key-value attributes and orderings, but it seems
unnecessary to add the infrastructure without a concrete need.

v3: - convert from comma to / as separator to avoid separator collision with
      other callers into name_to_dev_t
v2: - convert from +/-%u to ,PARTNROFF=%d

Signed-off-by: Will Drewry <wad@chromium.org>

TEST=booted with root=PARTUUID=%U/PARTNROFF=1 (in this case, root-a)
     tested with root=PARTUUID=<kern-c>/PARTNROFF=-1 (root-b)
     tested with root=PARTUUID=<kern-a>/PARTNROFF=3 (root-b)
     tested with root=PARTUUID=%U/PARTNROF=1 (malformed string caught)
BUG=chromium-os:14127

Change-Id: I05288c4c8373a2c7b44fda1e8ac097a2616f988e
Reviewed-on: http://gerrit.chromium.org/gerrit/1111
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>