cfg80211: Ignore downstream DEAUTH for authtry_bsses
If we receive a DEAUTH message from an AP while performing
an AUTH request, this does not constitute a reply to our AUTH
request.  In some cases an incoming DEAUTH may be due to the AP
clearing out old state from a previous authentication.
Instead we should ignore it and continue waiting for an AUTH
response.

Signed-off-by: Paul Stewart <pstew@chromium.org>

BUG=chromium-os:14765
TEST=New StrictAuth automated test (soon to be added to roaming suite)
MatFunc, SecMat

Change-Id: I8a58956d4c329d05556e1a563800a00cd0a5ba4a
Change-Id: I4306b66831320ad021d9f2b98fddedd1a0747b47
Reviewed-on: http://gerrit.chromium.org/gerrit/2186
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>