nvhost: Do not crash if wait syncpt > max
If user space is waiting for a syncpt value higher than max,
nvhost_wait_timeout() panics the kernel. This panic is based on the
assumption that nobody should wait for sync points higher than indicated
by a submit from user space.

Instead of causing the kernel to panic when a request from user space
hits this failing condition, fail the command so that it triggers
the appropriate error path inside the user space code where this
situation should properly be addressed.

TEST=verified that jsgamebench no longer causes kernel panic.
BUG=None

Change-Id: Ia714cfdbdace5ec87f5c9189429d8d9f20f27d20
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/4456
Reviewed-by: Olof Johansson <olofj@chromium.org>