CHROMIUM: ARM: tegra: change the clk_writel() macro from __raw_writel() to writel()
A write operation on PPSB, APB and AHB bus can be completed out-of-order.
Especially if the AHB or APB clocks are been diveded down or if there is heavy
traffic on the AHB or APB buses which can delay the writes on those bus. This
change can guarantee any in-flight MMIO write operations complete before we
disable or reset the clock on PPSB, APB or AHB bus.

BUG=chromeos-partner:2851
TEST=test it with enable DVFS/EMC_SCALING/CPU_IDLE and pass over weekend reboot loop test

Change-Id: If9a4d28b610aef081752457d098b12513ecc2276
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/2817
Reviewed-by: Grant Grundler <grundler@chromium.org>