Add bootstages to record timing in the kernel.
We can record timing in the kernel by invoking bootstage_mark() or
bootstage_mark_early for early stages before memory and timekeeping
are initialized. It also uses debugfs to interact with the user space.

BUG=chromium-os:12250
TEST=Manually tested by cat /sys/kernel/debug/bootstage/report to see
the result and writing to /sys/kernel/debug/bootstage/mark to record
timing, and see report again.

Change-Id: I92d410b9c48463c69d1d13ae02fe3f62f7bfd223
Signed-off-by: Da Zheng <zhengda@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/4892
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Da Zheng <zhengda@chromium.org>