CHROMIUMOS: mac80211: Drop DS Channel PARAM in directed probe
Do not send DS Channel parameter for directed probe requests
in order to maximize the chance that we get a response.  Some
badly-behaved APs don't respond when this parameter is included.

Signed-off-by: Paul Stewart <pstew@chromium.org>

BUG=chromium-os:16825
TEST=Manual: Test against misbehaving AP, ensure in logs that
connection was made due to a successful direct-probe.  Do packet
trace of background scan to make sure this still contains the
DS Channel Parameter EID.

Change-Id: I17fa91b6de8b2511f967bfd71ac4bc49b2408248
Reviewed-on: http://gerrit.chromium.org/gerrit/2995
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>