CHROMIUM: arm: tegra: clocks: sanity-check clocks at boot
This goes through the list of clocks, checking rates at the end of boot
and adjusting down any that are set beyond max, while also warning on
the console.

Scarily enough, there are actually several enabled clocks that are run way
out of spec with the current kernel. Fixing those will be done separately.

BUG=none
TEST=boot, check dmesg for output. Contemplate opening bugs for any errors printed.

Change-Id: I6ae017756abc0df3169fb57edae10c3c32085105
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/3391