CHROMIUM: arm: tegra: reparent most clocks from pll_m to pll_c
pll_m is used for memory, and varies between T20 and T25. pll_c is
kept at 600MHz on both parts, and is a better parent for some of the
miscellanous clocks on the chip.

BUG=none
TEST=manually look at /sys/kernel/debug/clock/clock_tree after this patch
and the 'purge clock table patch that goes after it' are applied, and make
sure that the only clock still under pll_m is emc.

Change-Id: Ie7c3c1586f70fe193e52fd33d2fd7d851ed0b14a
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/4292