CHROMIUM: config: tegra: Switch early printk to UARTB.
This will BREAK earlyprintk for seaboard but fix it for many
seaboard variants (which all use UARTB).  If we want to do something
better, we'd need to do something clever like store the printk UART
in a scratch register.

BUG=chromium-os:18486
TEST=On kaen, added earlyprintk to the cmdline and saw early printouts.
Note that you also see some garbage printed out before the early
printouts.  That seems fairly harmless, but should probably be fixed
eventually.

Change-Id: I9db422a3f3bd7122b6c9fcfafb55b66f97084544
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/5769
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>