CHROMIUM: Disable VM_MAYEXEC tainting for noexec mounts
With our userland, VM_MAYEXEC tainting does not provide much
additional benefit beyond protecting against LD_PRELOAD or
dlopen()ing files dropped in a noexec mountpoint.  This sets
the sysctl default to 0 such that VM_MAYEXEC is not masked off
of /dev/shm and other mountpoints when a file is mmap'd.

Signed-off-by: Will Drewry <wad@chromium.org>
BUG=chromium-os:19221,native-client:1883
TEST=booted, ran mmap_tester.c from the nacl bug.

Change-Id: If3c84f7c000b22328e8980fdf3cbcdb155a82a4b
Reviewed-on: http://gerrit.chromium.org/gerrit/6081
Reviewed-by: Mandeep Singh Baines <msb@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>